+++ src/duckdb/schemas/tracks.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS tracks (
-  id          BIGINT PRIMARY KEY,  
+  id          INTEGER PRIMARY KEY,  
   category_id INTEGER NOT NULL,    
-  video_id    BIGINT NOT NULL,     
+  video_id    INTEGER NOT NULL,     
   FOREIGN KEY (video_id) REFERENCES videos(id)
 );
